Issue-127: all implementations moved from RestlasticSearchClient to RestlasticSearchClient6
